Fix issues when setting ORBIT_API_ENABLED to 0
The primary issue, visible externally, was that setting `ORBIT_API_ENABLED` to 0
caused all the new macros ending in `AND_GROUP_ID` to not be defined.

The secondary issue was that some variables in Orbit code ended up being unused,
causing build errors, so I added some `[[maybe_unused]]` in a few places.

Bug: http://b/206594767

Test: Build with `ORBIT_API_ENABLED` set to 1 and build with it set to 0.