Fix bug in PE/COFF header parsing code where one field was missing
One field (nsects) was not parsed in the COFF header parsing method,
which lead to all subsequent parsed to be incorrect (as everything
was offset by 2 bytes). This is fixed here. A check on the optional
header size is added into the parsing code, which would have
uncovered this issue. While this is not checking precisly against what
was incorrect, the optional header size is one of the few data fields
in the header that we can directly validate in the parsing code itself.

Tested: Unit tests.
Bug: http://b/192514457