Don't mark as errors samples in injected library not from trampolines (#3061)
With https://github.com/google/orbit/pull/2985 we were using the fact that a
sample has at least one frame in `liborbituserspaceinstrumentation.so` as an
indication that the sample fell in code called by a trampoline (and would fail
to unwind anyway). That sample would be marked as of type
`kInUserSpaceInstrumentation`.

But that didn't consider that `ForwarderThread` (the thread that sends the data
to OrbitService) also runs code in that library. Those threads were reported as
`kInUserSpaceInstrumentation` but they are perfectly fine samples.

We now detect if a sample is in code called by a trampoline if it has at least
one frame in `liborbituserspaceinstrumentation.so` AND one preceding frame
(usually the outermost where unwinding stops) in a trampoline.

Also, in `--devmode` we now show the error type of a callstack in the tooltip in
the capture view (very useful for this kind of investigations).

Bug: http://b/206783021

Test:
- Update unit tests.
- Capture Trata with sampling and user space instrumentation and have a look at
  samples; in particular:
  - verify that samples in `ForwarderThread` are ok;
  - use tooltips to verify that the type of errors is
    `kInUserSpaceInstrumentation` where it makes sense;
- Capture custom target with sampling (both DWARF and frame pointer unwinding)
  and have a look at samples.