Make internal library path absolute (#3070)
A recent PR (#2984) introduced the requirement that the path returned by
GetLibraryPath needs to be absolute (which is a good thing). The
internal handling didn't take this into account, so this change ensures
that also in the internal case an absolute path is returned.

Unfortunately this absolute path will still be based on the current
working directory. It is possible to determine the location of the
library based on the test executable location but that changes between
test executables and can't be determined in general here.

The best solution is probably to make the list of checked paths
changeable by the tests instead of being hard-coded. But that's a little
more effort, hence this quick fix.