Take module into account in BuildCaptureOptions (#3072)
BuildCaptureOptions used to assume that the added functions are part of the
module containing the `main` function. That's not the case when the tests
are built using shared linking and the main module only contains GTest's
main function while all the test cases are located in shared libraries.

This commit fixes that by also returning the module file path when
looking for the function's address.