Add Microsoft's cppwin32 project to third_party
The cppwin32 project implements a C++ language projection for the
Windows SDK APIs. The plan is to repurpose the code generation parts
of cppwin32 for Orbit's Windows API tracing feature.

This is a snapshot of the cppwin32 repo at commit
8a6b2507734dd9e9892059b5794f35109688fc20.

The cppwin32 project includes a snapshot of github.com/microsoft/winmd,
a winmd parser written in C++.

Note that some files have been pruned out to minimize the footprint.
Those are mainly test .winmd files that we don't need since we'll have
our own version.