Introduce namespaced includes for protos (#3079)
This is making a long overdue change on how we include protos. So far
we included protos only by specifying the proto's base name:

```
 #include "capture.pb.h"
```

This commit changes all includes to `modulename/basename.pb.h`:

```
 #include "GrpcProtos/capture.pb.h"
```

or

```
 #include "ClientProtos/preset.pb.h"
```

I'm doing this change now because module-name-prefixed includes allow
copybara to rewrite them to absolute paths with a simple
string-replace-rule. This was not possible before.

It's a big change. The relevant non-trivial changes are in:
- `cmake/grpc_helper.cmake`
- `src/ClientProtos/CMakeLists.txt`
- `src/GrpcProtos/CMakeLists.txt`
- `src/MetricsUploader/CMakeLists.txt`
- `src/OrbitTriggerCaptureVulkanLayer/CMakeLists.txt`
- `src/MetricsUploader/orbit_log_event.proto` (This file was moved!)