Extend EventLoop functionality
This changes the EventLoop::exec functionality to not just allow and
error before its running, but also a non error result (integer value).

It also slightly tweaks the behaviour if Eventloop so it can be
restarted (exec called again), to resemble the behaviour of QEventLoop.

Part of http://b/174561221
Test: Unit test + starting & connecting of Orbit to manually check
that everything still works