Fix TracingHandler::ProcessFunction{Entry,Exit} after tracer_.reset() (#3082)
We leave the reset to the destructor.
I also tried the solution using a mutex, but the behavior at the end of a
capture with a very large number of event was different. In particular, a huge
gap in instrumentation data at the end of the capture could be observed. So I
think that this is a safer option.

Unfortunately I was never able to actually reproduce the crash so that makes it
harder to judge what the best solution is.

Bug: http://b/207480127

Test: I tried this on a synthetic benchmark with various functions to
instrument, including one called more than 10M times per second.