Add initial fuzzing for PE/COFF parsing in libunwindstack
dfhis adds a first fuzzer for PE/COFF parsing. The fuzzer is extremely
basic and generates very poor coverage as it is not at all aware of
the structure of PE/COFF files. This will be fixed later. For now,
this is added to establish the structure of how we add fuzzing tests
for libunwindstack and to make sure it shows up in our oss-fuzz
integration. Next steps will be to add a custom mutator and/or add a
corpus of PE/COFF files to trigger more code paths.

Tested: Ran fuzzer locally.
Bug: http://b/192514457