Fix crash in recursive instrumentation.
The previous "fix" restored the return address in the entry payload. But
the return address was altered *after* the call to the entry payload. So
this never worked.

Removed the overwriting of the return address from the trampoline and
shifted it into the entry payload. This way we can decide if it is
appropriate to do the instrumentation or not depending on wether we are
re-entering the payload.

Bug: http://b/207361417
Test: Ran locally. Multiple times.