Add fuzzing for 32-bit PE/COFF files
The initial fuzzer implementation for PE/COFF only worked on 64-bit
files, this is now fixed to also include 32-bit.

I have also removed a comment that says that the basic fuzzer
generated poor coverage because the coverage on oss-fuzz actually
is quite good.

It is unclear to me whether the 32-bit and 64-bit versions should
run in the same fuzzer target or whether we should have two targets.
The current version (both in the same fuzzer) is favored so far,
as the corpus (which is generated incrementally by oss-fuzz) is
actually quite similar for 32-bit and 64-bit, so this should benefit
coverage. In addition, reading the code coverage report with both
versions in the same fuzzer is easier: Without one version, the
coverage reports contains large parts indicating that the other
version wasn't instantiated.

Tested: Built & ran fuzzer locally.
Bug: http://b/208200278