Remove usage of internal gRPC API
Everything in the `grpc_impl` namespace is not part of the public API
and prone to change.

In this particular case `grpc_impl::Channel` can simply be replaced by
`grpc::Channel`.

I discovered this while trying to update gRPC to a newer version which -
of course - made changes to this part of the code.