Remove update flag from conan calls (#3105)
The `--update` or `-u` flag to conan commands makes conan check the
remote servers for updated recipes and packages. If updated ones
are found they will be downloaded. But that makes no sense in our
case because we use lockfiles which do not allow updating the packages
anyway. Nevertheless the flag has been in-place to work around a bug in
a previous conan version which has been fixed for a while now.

So this commit removes the flag which should also speed up the build a
bit because it avoids rountrips to the server.

I'm changing this now because I have the suspicion that this might be
related to our failing tests on the CI.