Mark dependencies as overrides in conanfile
If a package appears multiple times in the dependency graph
with different versions, conan forces the root package to
take a stance and choose one - the correct - version of the
dependency.

That's why Orbit has a couple of `self.requires` calls in its
conanfile.py for dependencies it doesn't directly use.

But these so-called overrides clutter the dependency graph when
serialized into the lockfile and make it harder for the user
to distinguish between real dependencies and mere overrides.

That's why the `self.requires` call takes a flag `override=True`.
It marks a dependency as override-only and as such keeps the
dependency graph clean and let others know that this package is not
directly needed.