Erase a CHECK in UpdateVisibleTrackList
Solving this crash: http://b/208391403.
This CHECK doesn't have a guarantee to be true, since a track could be
added between SortTracks and UpdateVisibleTrackList. Another option is
to lock that method but it's not needed, because we still can insert new
Tracks to be sorted in the next Draw call.

This is still an issue because we have two threads running in
TrackManager (http://b/177200020), and we are aiming to solve this by
making TrackManager to only work with main_thread.