Add E2ETestCase's for capture loading
This change adds an E2ETestCase that attemps to load a capture of
a given file. It waits until the main window opens.

Further, it adds an E2ETestCase to verify that a model window exists and
to click on the "OK" button. This will be used, to verify loading of
old captures will raise an error.

Test: Wrote an E2E test and run it locally (will be a different PR)
Bug: http://b/192222100