Clean shutdown (#2913)
* Refactor ServiceDeployManager StopSftpChannel

Remove StopSftpChannel helper function, rename StopSftpChannel to
ShutdownSftpChannel and add nullptr checking and resetting of
sftp_channel_ in ServiceDeployManager::Shutdown

* Improve OrbitSshQt::Tunnel

Adds missing wait for remote shutdown step.

* Fix problem with tunnel stop

Tunnel::Stop can ne called from 2 different locations. 1. When the
local_socket_ signals a disconnect and 2. from ServiceDeployManager
ShutdownTunnel. If 1. happens before 2., ShutdownTunnel will wait
forever for a stopped signal. This is here solved by emitting stopped()
when the Tunnel is already stopped.

* Fix Bug in ServiceDeployManager::ShutdownTunnel

In ShutdownTunnel tunnel the wrong signal (started instead of stopped)
was used to quit the eventloop

* Refactor ServiceDeployManager::ShutdownTunnel

Refactor the method signature to be similar to other shutdown methods
and move optional checking and setting of grpc_channel_ into
ServiceDeployManager::Shutdown

* Add watchdog stopping to Shutdown

This stops the watchdog in ServiceDeployManager::Shutdown and adds a
CHECK to the watchdog start.

* Fix bug in OrbitSshQt::Task

There was a `break` in Task::shutdown(), which resulted in shutdown
returning success after sending EOF instead of waiting for the answer of
the remote side.

* Add timeout to Task::Shutdown

* Refactor ServiceDeployManager ShutdownOrbitService

Refactor the method signature to be similar to other shutdown methods
and move optional checking and setting of orbit_service_task_ into
ServiceDeployManager::Shutdown

* Refactor ServiceDeployManager::ShutdownSession

Refactor the method signature to be similar to other shutdown methods
and move optional checking and setting of session_ into
ServiceDeployManager::Shutdown

* Run Shutdown in ServiceDeployManager dtor

http://b/b/174561221