Add dll injection to WindowsUtils
To inject a dll into a remote process, we first remotely allocate memory in which we write the
name of the dll to be loaded. We then create a remote thread which calls LoadLibraryA from
kernel32.dll with the address of the previously allocated string as parameter. We also provide
GetRemoteProcAddress to find the address of a function in a remote module.