Timegraph interface (#3120)
This moves all read-only functionality related to time / world conversion
into a separate interface and passes this to Track. This removes the
dependency to TimeGraph from most of OrbitGl.

The TrackManager still has a TimeGraph dependency as it requires
TimeGraph::VerticallyMoveIntoView. We can get rid of this in the future.