Add E2E test for loading a capture (#3135)
This adds an E2E test that verifies loading a capture (and failing on
old captures). For a successfully load, it also verifies various tracks
and data to be correct.

This adds two capture files of OrbitTest taken with versions 1.64 and
1.72. The files are added to Github, such that the test can be updated
atomically, allowing the internal infra to also execute the previous
version for compatbility verification.

In order to use capture files from Github, the LoadCapture test was
adjusted to also support relative paths. Further, a utility function
to retrieve the orbit base path was added.

Test: Manually store the two capture files into a nightly Orbit
      directory and run the test.
Bug: http://b/192222100