Rename temporary Orbit.zip to OrbitUI.zip
Our CI system compiles collector and UI in separate jobs and combines
both artifacts into a single ZIP archive called Orbit.zip.

It can come to problems when both the resulting artifact of the UI
compilation job and of the combiner job are both called Orbit.zip.

This fixes the glitch by renaming the output filename of the UI
compilation job.