Ensure dialogs get closed when the session ends
This applies to both the source code and the disassembly view dialog
windows.

I also tried making the widgets children of the main window (the natural
thing to do in Qt-land), but that messes with the modal status of the
windows. This could probably be fixed by changing the window flags, but
just using a signal-slot connection is simpler and less brittle.

Bug: http://b/208956231
Test: Tried manually