Update Qt to version 5.15.2 (#3130)
This transitions from our custom Qt 5.15.1 recipe to the official Qt
5.15.2 recipe from Conan Center. That reduces the recipe maintenance
burden and also saves us from manually updating all the dependencies
away from the deprecated bincrafters packages.

With this change I also removed the QtWebEngine support because that
saves us a couple of hours in compilation time (per build). It's a lot
more efficient to just enable it again when we need it instead of
keeping it around.