Extend E2E test to also load latest capture
This adds additional steps to the capture loading E2E test, that verify
taking a capture automatically saves a capture and that this capture
shows up in the list of latest captures.

To do so, the test takes a capture of hello_ggp and afterwards filters
in the recent captures list for the target name and the current date.
And loads the most recent match.

Note, as in theory the test could be run arround midnight, and the
capture might get actually taken on the next day, the test also checks
for captures on the next day (if non is found for today).

Test: Start hello_ggp and run the E2E test locally.
Bug: http://b/192222100