Remove active waiting from the CopyFileToLocal routine
Active waiting means the waiting function will not end but rather
execute other waiting events on the event loop.

But this will increase the stack size with every actively waiting
function and eventually leads to a stack overflow.

So this commit removes active waiting and instead returns a future. That
means the caller needs to take care of waiting. In this case waiting is
actually not needed and the caller will just chain up a couple of
futures and be done.