Update to orbit_api_start_async_v1 (and orbit_api_v2) (#3153)
This brings `orbit_api_start_async` in line with `orbit_api_start_v1`, where we
can distinguish between the program counter being passed from the outside or it
being retrieved as the function's caller (when `kOrbitCallerAddressAuto` is
passed).

There are no changes in how `Orbit.h` and its macros are used.

This will make things a bit cleaner in the implementation of Orbit API for Wine,
where functions for one platform forward calls to functions for the other
platform.
But consistency with `orbit_api_start_v1` is also a desired effect.

The version numbers of `struct orbit_api_v#` and of the individual functions are
kept independent. Viceversa, names of all functions would have to be changed for
every version bump of `struct orbit_api_v#`. A similar example is LibGGP.

Test: Tested that OrbitTest compiled with `orbit_api_v1` still produces correct
data, in particular regarding async scopes. Tested that OrbitTest with the new
version produces correct data. Also tested as part of the prototype for Orbit
API on Wine.