Define PeCoffInterface base class to be in line with ELF code (#3159)
This is in preparation for properly integrating the PE/COFF unwinding
by adadpting the PeCoffInterface interface to match what is required,
mainly so that the PE/COFF implementation can simply follow the ELF
implementation. The main changes are to separate out a base class to
allow interfacing with both the 32-bit and 64-bit case, and to add
a load_bias in/out parameter to Init, following the ELF implementation.
(Note that using the image_base value for load_bias is not a precise
one-to-one mapping, but works for our purposes and our address
computations. I'm not sure this is the best approach, but it works.)

Tested: Unit tests.
Bug: http://b/192514457