Implement PE/COFF equivalent of Elf class in libunwindstack
This change implements the class PeCoff, which is the PE/COFF equivalent
of the Elf class. Its main reponsibility is to correctly initialize
the PeCoffInterface based on the type of PE/COFF file (32-bit vs.
64-bit) and pass through calls to the PeCoffInterface.

To reuse fake PE/COFF test data for PeCoffTest, the existing
PeCoffInterfaceTest is refactored by extracting the PeCoffFake class
that can be initialized with fake PE/COFF data and also allows
manipulation of the fake data to trigger failure cases in the
parsing code.

Tested: Unit tests.
Bug: http://b/192514457