Detach Unwinding Method from Sampling
Before this change, "no sampling" was represented by
setting the unwinding method to "undefined". However,
in the future, we also want to do callstack collection
and unwinding on different triggers (such as uprobes).
Therefore, we want to set the unwinding method even
when not sampling. We use sampling period of "0" as
a new way to disable sampling.

Further, this change reduces the default stack dump
collection size when frame pointer unwinding.
std::numeric_limits<uint16_t>::max() is used
when the stack size is not defined, and the new
defaults should be used.

Test:
Take captures with and without frame pointers.
Run Tests.