Remove active waiting from background thread (#3165)
With removing the call to `DeferToBackgroundThreadAndWait` from
`CopyFileToLocal` the stack overflow on the main thread has been
mitigated. But now we process many file copies and parallel in a
background thread that is doing the actual work. Hence, the stack
overflow will appear there and the problem just shifted.

So this commit also removes active waiting from the background thread
function `CopyFileToLocalImpl` and replaces it by a Future-based
approach.

That's not super simple and unfortunately the code is harder than I had
liked. I tried a couple of different approaches and I believe this is
the simplest one, while not considering larger refactorings for now. Of
course the best approach would be to rewrite `SftpCopyOperation` based
on Futures, but that's quite some effort and not a quick fix.

This change also surfaced another problem. Previously we have been
opening a separate SftpChannel for each file copy operation which
becomes a problem when the number of copy operations exceeds the max
number of channels. I wrote it like that because I thought
libssh2 has a bug in handling mutiple files over the same channel.

But while revisiting that I realized that my assumptions were wrong
and we were using libssh2 in a wrong way. I created a bug for that
(http://b//209807583), but until that's fixed properly we still need
to work around the problem.

So I introduced a queue for all the file copy operations such that they
will processed all one after each other instead of in parallel.

To be sure that it works this time I created a simple application with
a hundred shared libraries and tried to load all of them in one click on
`Load Symbols`. Works fine. Some tests with Trata also worked fine.

Bug: http://b/209461845