Fix status message when downloading symbols (#3167)
This fixes a regression that has been introduced by PR #3157.

Since the SecureCopyCallback is not blocking anymore, showing the
download message while this function is executing is pointless. It is
fixed by moving the status handler into the completion closure.

The SCOPED_TIMER_LOG macro has been replaced by two dedicated log
messages. Unfortunately the macro does not allow the underlying type to
be moved somewhere else. To support that a bigger restructuring of all
the logging methods is needed and would exceed the scope of this PR.
Hence the work around with the two log messages.