FIx instance connection bug
Beforehand, when the user started a connection to one instance and
while the call to GetSshInfoAsync was running selected a different
instance and connect to it, Orbit would crash. This happened because
when the first GetSshInfoAsync returned,
ConnectToStadiaWidget::OnSshInfoLoaded was fully executed and did not
take into account that the selected instance could have changed in the
mean time. This is now fixed

Test: Start connection to instance, immediately cancel and start a
connection to a different instance.

http://b/209430645