Update git-describe arguments (#3179)
Due to http://go/stadia-orbit-version-numbers we have to make some
changes to how we call `git describe`. We have to remove the `--tags`
option such that `git describe` will ignore lightweight tags and only
consider annotated tags.

Furthermore we want to add the `--dirty` flag such that builds with
changes in the local worktree are marked accordingly.