Port OrbitService to Windows
Add "Windows" and "Linux" directories in "src/Service" which
contain platform-specific implementations of OrbitGrpcServer
and OrbitService. Also, compile ProducerSideChannel on Windows
and extract Linux specific code from ProducerSideServer.

Rename ServiceLib -> OrbitServiceLib so that it appears next to
the OrbitService project in VS.

Tests: Run OrbitService on Windows and Orbit with "--local".
We can now take a capture and Scheduling and sampling events
are now piped to the client.