Prevent long repeated strings with repeated ORBIT_ASYNC_STRINGs
When we used `ApiEvent` to carry manual instrumentation data, we allowed long
strings by concatenation. This is no longer the case and can lead to a very long
repeated string if `ORBIT_ASYNC_STRING` is called multiple times with the same
id. This is somewhat easy as the API is not too easy to use, and can also tank
the performance of the client. Let's handle this case gracefully, while still
allowing concatenation for the old format.

Bug: http://b/207483525

Test:
- Put together a binary that sends async scopes with a string reusing the
  same id. Capture it and verify the string displayed in the capture view is now
  not repeated.
- Take a capture on the same binary using Orbit.h 1.66 with an Orbit 1.66, using
  `ApiEvent` for manual instrumentation. Load it with the current version of
  Orbit and verify the string is still repeated.