Merge directory automation_tests/test_data into testdata
We have two "test data" directories in `automation_tests`, which is a bit
confusing. Let's keep the older one and move the content of the newer one into
the older one.
We keep the older one because it's a much simpler change; in particular, we
don't need to modify Go code.

Test: Ran orbit_capture_loading.py locally.