Add clang11 docker containers
This is adding a Dockerfile for Ubuntu-based containers that come with
the clang-11 compiler.

Clang-11 is only part of Ubuntu starting with version 21.04. But Ubuntu
21.04 also ships with glibc 2.33 which is not compatible with the
comparatively old Linux kernel version on our host VMs.

That's why this container is based off of Ubuntu 20.04 LTS (as our other
containers) with clang-11 from the apt.llvm.org repository.