LinuxCaptureService stops the capture when it detects high memory usage
A thread polls the resident set size during the capture and verifies it stays
below half the total physical memory. I just picked a simple threshold, we can
adjust the value.

The client needs basically no change. `CaptureClient` will simply transition
from `State::kStarted` to `State::kStopped`. That is, unless the client also
stops the capture after the service has interrupted it but before
`CaptureFinished` has been sent; but this also works as expected.

The new `CaptureFinished::Status::kInterruptedByService` is added to
`CaptureFinished`, to signal that the capture was stopped by the service and not
by the client.

The trickier part was to make this work with the streaming Capture gRPC. In the
synchronous version that we use, when we stop the capture on the service side,
and hence the client doesn't call `WritesDone`, `ServerReaderWriter::Read`
blocks until the RPC finishes, i.e., until `CaptureService::Service::Capture`
returns. So we need to move the `Read` to a different thread and join it later.
I also tried switching to a callback-based async gRPC service, but the code was
more complicated, and it would require updating to gRPC v1.39.0, which is
complex.

Bug: http://b/209364229

Test: Taken a large number of captures instrumenting functions that caused
OrbitService to be unable to send data quickly enough, hence eventually hitting
the memory threshold. These cases saw OrbitService being killed before this
change.