Add DataView::Init() method
The "sorting_column_" member was not properly initialized which
resulted in first calls to "OnDataChanged" to always sort the
DataView by the first column. We can't initialize "sorting_column_"
in the "DataView" constructor since it requires a virtual function call.
We therefore introduce an "Init" method to be called after construction
of a "DataView".

More info can be found in the bug below.

Bug: http://b/211912871
Test: Multiple captures and manual sorts of data views.