Isolate the timeline into a proper class
In this PR, we are creating a proper class for the TimelineUI.

The class diagram could be found in this doc:
http://go/stadia-orbit-timeline-improvements#bookmark=id.yl5zta1faen2.

The Timeline class will be added as a children of the TimeGraph
after splitting the proper TimeGraph with a new TrackContainer class.

We needed to add a method into TimelineInfoInterface because we want
nanosecond precision with Ticks.

TODO: next commit will be the accesibility class for the timeline.

In http://screen/6NpHsWJECytpxuf we can see how ticks will look like
(top one) and the current timeline (bottom one).

Bug: http://b/208431620