Simplify vertical zoom method
We are refactoring TimeGraph to also include the TimelineUI component.
This means that Timeline should react to Vertical Zoom as well. As it
was difficult to understand the part of proposed-ratio/real-ratio, I'm
simplifying here this method to make the refactor simpler later.

I hope that now this method will be easier to understand.

Test: Load a capture, test vertical zoom.