Disconnect bitness of DWARF and target architecture in libunwindstack
This change disconnects the bitness of DWARF (DWARF-32 or DWARF-64)
from the bitness of the target architecture, which were (incorrectly)
connected in libunwindstack when no encoding was given explicitly. The
main change is only a few lines of code in DwarfSection.cpp that set the
encoding size to the size of the target address size when needed (this
is potentially overridden later by explicit choice in the augmentation string
or by the address size parameter in DWARF version 4 or 5).

Relevant sections in the DWARF standard 4 and 5 (available
at https://dwarfstd.org/) are 7.4 and 7.5.2.

While the actual code change is small, multiple unit tests assumed the
original behavior and needed to be adapted. In detail:

* Fixed PeCoffInterfaceTest due to incorrect CIE/FDE setup (wrong type
sizes for the target address).
* Fixed DwarfEhFrameTest by enforcing encoding size using augmentation
string. The current code for .eh_frame parsing assumes a value size of
4 bytes for encoded target addresses. This is probably okay in practice,
as the encoding is typically (always?) given explicitly in the .eh_frame
case, but we need to work around this for unit testing.
* Fixed DwarfEhFrameWithHdrTest by enforcing encoding size using
augmentation string, for the same reason as stated above.
* Fixed DwarfDebugFrameTest by correcting the FDE setup and making it
depend on the chosen target address size.