Allow DWARF column numbers beyond actual registers in libunwindstack
DWARF allows column numbers for the DWARF table for unwinding frames
that are larger than actual (general purpose) register numbers
(sometimes these are called virtual registers). So far, libunwindstack
would reject such column numbers. We now allow column numbers up to a
certain limit by expanding the vector that is storing the registers
during evaluation of DWARF expressions when unwinding a frame.

Note that an alternative would be to simply increase the maximum
number of registers present for all architectures (or the architectures
we care about), and this was also how this was initially done in our
unwinding prototype. However, the "column number larger than general
purpose registers" feature is a feature of DWARF and thus should be
handled and controlled by code related to DWARF.

A possible (but potentially larger) refactoring would be to disconnect
the columns in the evaluated DWARF table from the registers (including
copying the subset of the table that belongs to actual registers into
the registers vector after evaluation).

Tested: Unit test added.
Bug: http://b/194768602