Add class for handling PE/COFF runtime functions in libunwindstack
This change adds a class for parsing, storing, and finding PE/COFF
runtime functions (as stored in RUNTIME_FUNCTION structs in the PE/COFF
file). These structs represent the address ranges of functions in a
module, can be found in the .pdata section of a PE/COFF file, and
include a pointer to the unwind information for the represented
function.

It is critical to be able to retrieve a runtime function entry quickly
during an unwind step to look up the corresponding unwind information,
hence we add a binary search on the (guaranteed to be sorted) list of
runtime functions entries.

See https://docs.microsoft.com/en-us/cpp/build/exception-handling-x64
for more on the RUNTIME_FUNCTION struct.

Tested: Unit tests.
Bug: http://b/194768602