Add class for handling PE/COFF unwind infos in libunwindstack
We add a class PeCoffUnwindInfos that is used to parse and cache
unwind infos (as defined by the UNWIND_INFO struct) from PE/COFF object
files.

It is similar to the corresponding class for handling runtime
functions (RUNTIME_FUNCTION struct), with some minor differences: (1)
The section bounds are not passed in, as they are generally not known
for  unwind infos (runtime functions are always in the .pdata section,
we don't know this for unwind infos), (2) not all data is parsed
upfront, we parse on request and cache the entry (based on performance
tests in the prototype, this is the better approach here).

Some changes are made to PeCoffRuntimeFunctions* files for consistency
or because I noticed some smaller issues (like using 'explicit' on the
constructor and more info on errors).

Tested: Unit tests.
Bug: http://b/194768602