Remove DataManager::selected_callstack_events_by_thread_id_
This was basically a duplication of
`CaptureData::selection_callstack_data_.callstack_events_by_tid_`.

Note that I also wanted to move `CaptureData::selection_callstack_data_` to the
`DataManager`, because the selection is not part of the capture.
The problem is that `CallstackData` is synchronized, so having it as a member of
`DataManager` clashes with the idea that everything in `DataManager` should be
accessed from the main thread, even if that member would indeed only have been
accessed from the main thread. This needs to be done in more steps, keeping only
`CallstackEvent`s (related to the selection) in `DataManager`.

Bug: http://b/214231145

Test: Capture Trata, make a few selections.