Handling of unwind operations for PE/COFF unwinding in libunwindstack
This is the first part to add native PE/COFF unwinding in
libunwindstack, and adds handling of all required unwind operations.

Documentation of the unwind operations can be found here:
https://docs.microsoft.com/en-us/cpp/build/exception-handling-x64?view=msvc-170

Unit testing is added to cover all operations, failure cases, and some
mixed cases with multiple operations. For testing, a call frame and the
corresponding unwind info is built so that we can verify that registers
are restored with the right values as read from the call frame on the
stack.

Tested: Unit tests.
Bug: http://b/194768602