Move Function/Module Lookup out of CaptureData
This change allows function and module lookup from
places, where we don't have a capture available.
To do so, it moves the lookup functions from `CaptureData`
to a new set of free functions.

Further, this moves `ModuleManager` out of `CaptureData`
and store it explicitly.

The process and module information, as well as if needed
the capture information, now need to be passed explicitly
to the new lookup functions.

An immediate side effect of this change, is that
`CaptureData` becomes a pure data class again.

Test: Run Tests
Bug: http://b/216075140