Show call targets in disassembly view (if possible)
Currently, in disassembly view, call instructions only show
a target address. In is quite some figure out what the actual
target function will be.

This change, does a simple approache to show the target function
name. It tries to convert the operant of a call instruction
into a hex number and check if there is a function under this
number interpreted as absolute address. If not it will show
[???] as function name (as in the other views).

As a result, the change will not be able to resolve relative
calls, as well as more complex calls (e.g. based on registers).

In general, function names will always be in addtion to the
actual disassembly string, i.e.:

call         0x401030 (int fib(int))

Test: Open disassembly from modules/function tab as well as from
      sampling tab. Also open the CodeViewerDemo (for highlighting).
      Run unit test.
Bug: http://177827187