Don't store PostProcessedSamplingData in SamplingReport
For the entire capture, `SamplingReport::post_processed_sampling_data_` was a
duplication of `CaptureData::post_processed_sampling_data_`. Store a pointer
instead.

For the selection, `SamplingReport::post_processed_sampling_data_` was the only
place where the selection's `PostProcessedSamplingData` was stored, which is
undesired as `SamplingReport` is a UI class. Store it in
`CaptureData::selection_post_processed_sampling_data_`. While this should
eventually be moved to the `DataManager` together with
`CaptureData::selection_callstack_data_`, it is desired that the two are now
stored together as they are closely related (one is derived from the other at
the end of the capture).

`SamplingReport::unique_callstacks_` was also a duplication of data from
`CaptureData::callstack_data_` and from
`CaptureData::selection_callstack_data_`, so we remove that too and keep a
pointer directly to the `CaptureData` instead.

Test: Take various captures. Make selections. Try capture loading.